feat: also initialize missing slots in sw-category-detail
